requirements.txt
python-keystoneclient
python-glanceclient
python-novaclient
python-quantumclient
python-cinderclient
python-swiftclient
python-neutronclient

